+++ src/main/java/com/actiontech/dble/server/handler/SetCallBack.java
@@ -6,17 +6,17 @@
 package com.actiontech.dble.server.handler;
 
 import com.actiontech.dble.server.variables.MysqlVariable;
-import com.actiontech.dble.services.MySQLVariablesService;
+import com.actiontech.dble.services.BusinessService;
 import com.actiontech.dble.sqlengine.SQLQueryResult;
 import com.actiontech.dble.sqlengine.SQLQueryResultListener;
 
 import java.util.Map;
 
 public class SetCallBack implements SQLQueryResultListener<SQLQueryResult<Map<String, String>>> {
-    private final MySQLVariablesService service;
+    private final BusinessService service;
     private final MysqlVariable[] items;
 
-    SetCallBack(MySQLVariablesService service, MysqlVariable[] items) {
+    SetCallBack(BusinessService service, MysqlVariable[] items) {
         this.service = service;
         this.items = items;
