+++ src/main/java/com/actiontech/dble/services/manager/information/tables/SessionVariables.java
@@ -7,7 +7,7 @@ import com.actiontech.dble.net.IOProcessor;
 import com.actiontech.dble.net.service.AbstractService;
 import com.actiontech.dble.server.variables.MysqlVariable;
 import com.actiontech.dble.server.variables.VariableType;
-import com.actiontech.dble.services.MySQLVariablesService;
+import com.actiontech.dble.services.VariablesService;
 import com.actiontech.dble.services.manager.information.ManagerBaseTable;
 import com.google.common.collect.Maps;
 
@@ -52,7 +52,7 @@ public class SessionVariables extends ManagerBaseTable {
                     forEach(fc -> {
                         AbstractService service = fc.getService();
                         if (service != null) {
-                            for (MysqlVariable var : ((MySQLVariablesService) fc.getService()).getAllVars()) {
+                            for (MysqlVariable var : ((VariablesService) fc.getService()).getAllVars()) {
                                 LinkedHashMap<String, String> row = Maps.newLinkedHashMap();
                                 row.put(COLUMN_FRONT_ID, fc.getId() + "");
                                 row.put(COLUMN_VAR_NAME, var.getName());
