+++ src/main/java/com/actiontech/dble/services/manager/response/ShowConnection.java
@@ -13,7 +13,7 @@ import com.actiontech.dble.net.IOProcessor;
 import com.actiontech.dble.net.connection.FrontendConnection;
 import com.actiontech.dble.net.mysql.*;
 import com.actiontech.dble.route.factory.RouteStrategyFactory;
-import com.actiontech.dble.services.MySQLVariablesService;
+import com.actiontech.dble.services.VariablesService;
 import com.actiontech.dble.services.manager.ManagerService;
 import com.actiontech.dble.services.mysqlsharding.ShardingService;
 import com.actiontech.dble.util.IntegerUtil;
@@ -246,8 +246,8 @@ public final class ShowConnection {
         }
         row.add(txLevel.getBytes());
         row.add(autocommit.getBytes());
-        row.add(StringUtil.encode(((MySQLVariablesService) c.getService()).getStringOfSysVariables(), charset));
-        row.add(StringUtil.encode(((MySQLVariablesService) c.getService()).getStringOfUsrVariables(), charset));
+        row.add(StringUtil.encode(((VariablesService) c.getService()).getStringOfSysVariables(), charset));
+        row.add(StringUtil.encode(((VariablesService) c.getService()).getStringOfUsrVariables(), charset));
         return row;
     }
