+++ src/main/java/com/actiontech/dble/sqlengine/SetTestJob.java
@@ -18,7 +18,7 @@ import com.actiontech.dble.net.mysql.ResetConnectionPacket;
 import com.actiontech.dble.net.mysql.RowDataPacket;
 import com.actiontech.dble.net.service.AbstractService;
 import com.actiontech.dble.server.variables.MysqlVariable;
-import com.actiontech.dble.services.MySQLVariablesService;
+import com.actiontech.dble.services.BusinessService;
 import com.actiontech.dble.services.mysqlsharding.MySQLResponseService;
 import com.actiontech.dble.services.mysqlsharding.ShardingService;
 import com.actiontech.dble.services.rwsplit.RWSplitService;
@@ -35,13 +35,14 @@ public class SetTestJob implements ResponseHandler, Runnable {
 
     private final String sql;
     private final SQLJobHandler jobHandler;
-    private final MySQLVariablesService frontService;
+    private final BusinessService frontService;
     private final MysqlVariable[] setItems;
     private volatile List<FieldPacket> fieldPackets;
     private final int userVariableSize;
     private final AtomicBoolean hasReturn;
 
-    public SetTestJob(String sql, SQLJobHandler jobHandler, MysqlVariable[] setItems, int userVariableSize, MySQLVariablesService frontService) {
+    public SetTestJob(String sql, SQLJobHandler jobHandler, MysqlVariable[] setItems, int userVariableSize, BusinessService frontService) {
+
         this.sql = sql;
         this.jobHandler = jobHandler;
         this.frontService = frontService;
